LayoutTests/http/tests/loading/cross-origin-XHR-willLoadRequest.html

<script>
    layoutTestController.dumpAsText();
    layoutTestController.addURLToRedirect("http:/localhost:8000/loading/resources/bar.txt", "http://localhost:8000/loading/resources/foo.txt");

    var request = new XMLHttpRequest();
    request.open("GET", "http:/localhost:8000/loading/resources/bar.txt");
    request.send();
</script>
<p>
    The console message above should report failure to load foo.txt due to cross-origin access, not a network error.
</p>
